requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.30.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
docstring_parser==0.16
e2b==0.17.1
e2b-code-interpreter==0.0.10
eezo==0.3.7
exceptiongroup==1.2.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
googleapis-common-protos==1.63.2
griffe==0.36.9
groq==0.9.0
grpcio==1.64.1
guardrails-ai==0.4.5
guardrails-api-client==0.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
instructor==1.3.4
jiter==0.4.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonrpcclient==4.0.3
jwt==1.3.1
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.12
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langfuse==2.38.1
langgraph==0.1.5
langsmith==0.1.84
lxml==4.9.4
markdown-it-py==3.0.0
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
nylas==6.3.0
openai==1.35.10
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
orjson==3.10.6
packaging==23.2
pinecone-client==4.1.2
pinecone-plugin-interface==0.0.7
protobuf==4.25.3
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydash==7.0.7
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rstr==3.2.2
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
watchdog==4.0.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
